requirements.txt
Flask==3.0.3
Flask_Login==0.6.3
flask_sqlalchemy==3.1.1
openai==1.33.0
SQLAlchemy==2.0.30
Werkzeug==3.0.3
PyYAML==6.0.1

gunicorn==20.1.0
psycopg2==2.9.9